+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - #267 - Styling inconsistency
 - #242 - Support setting the max bytes to fetch per request
 - #247 - Support SCRAM once released
+- #271 - Provide an after_init option to pass a configuration block
 - Lower ActiveSupport dependency
 - Remove configurators in favor of the after_init block configurator